requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
openai>=1.0.0
python-dotenv==0.21.0
requests>=2.31.0
gunicorn==20.1.0
Pillow==10.0.0
pytesseract==0.3.10